requirements.txt
django
Pillow
djangorestframework
drf-yasg